chore: Reduce python requirements
Currently trying to build worker and api into python 3.11.

To update SQLAlchemy from 1.3 to 1.4 they dropped support for the schema
`postgres://`, now only `postgresql://` is supported.
Changing default value for DB because of that.

Also relaxing the constraint for setup.py and changing the image to be python 3.11 as well.